+++ src/main/java/skylark/command/CommandList.java
@@ -1,7 +1,7 @@
-package skylark;
+package skylark.command;
 
 /** List of commands available for the user. */
-enum CommandList {
+public enum CommandList {
     COMMAND_BYE, COMMAND_LIST, COMMAND_DONE,
     COMMAND_UNDONE, COMMAND_DEADLINE,
     COMMAND_TODO, COMMAND_EVENT, COMMAND_DELETE, COMMAND_FIND;
