+++ src/test/java/skylark/task/DeadlineTest.java
@@ -1,16 +1,18 @@
-package skylark;
+package skylark.task;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import org.junit.jupiter.api.Test;
 
-public class EventTest {
+import skylark.main.SkylarkException;
+
+public class DeadlineTest {
 
     @Test
     public void instantiate_dateTime_exceptionThrown() {
         try {
-            new Event("Make tea", "Invalid time");
+            new Deadline("Make tea", "Invalid time");
             fail(); // the test should not reach this line
         } catch (SkylarkException e) {
             assertEquals("Cannot parse date", e.getMessage());
@@ -19,15 +21,15 @@ public class EventTest {
 
     @Test
     public void toString_dateTime_success() throws SkylarkException {
-        Event task = new Event("Make tea", "2019-10-15 1800");
-        assertEquals("[E] [ ] Make tea (at: Oct 15 2019)", task.toString());
+        Deadline task = new Deadline("Make tea", "2019-10-15 1800");
+        assertEquals("[D] [ ] Make tea (by: Oct 15 2019)", task.toString());
     }
 
     @Test
     public void toStringFile_dateTime_success() throws SkylarkException {
-        Event task = new Event("Make tea", "2019-10-15 1800");
-        assertEquals("E | 0 | Make tea | 2019-10-15 1800", task.toStringFile());
+        Deadline task = new Deadline("Make tea", "2019-10-15 1800");
+        assertEquals("D | 0 | Make tea | 2019-10-15 1800", task.toStringFile());
         task.markAsDone();
-        assertEquals("E | 1 | Make tea | 2019-10-15 1800", task.toStringFile());
+        assertEquals("D | 1 | Make tea | 2019-10-15 1800", task.toStringFile());
     }
 }
