+++ src/test/java/skylark/task/TodoTest.java
@@ -1,4 +1,4 @@
-package skylark;
+package skylark.task;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
