+++ src/test/java/skylark/utils/PrinterTest.java
@@ -1,4 +1,4 @@
-package skylark;
+package skylark.utils;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
